+++ juniper-from-schema/tests/compile_pass/naive_date_time.rs
@@ -9,10 +9,10 @@ juniper_from_schema::graphql_schema! {
     }
 
     type Query {
-        dateTime: DateTime! @juniper(ownership: "owned")
+        dateTime: DateTimeUtc! @juniper(ownership: "owned")
     }
 
-    scalar DateTime @juniper(with_time_zone: false)
+    scalar DateTimeUtc @juniper(with_time_zone: false)
 }
 
 pub struct Query;